+++ lib/Plugin.php
@@ -155,7 +155,9 @@ class Plugin {
 			return $vars;
 		});
 
-		\add_action( 'init', array( $this, 'text_url_permalink' ), 99 );
+		\add_filter( 'post_rewrite_rules', array( $this, 'rewrite_rules' ) );
+		\add_filter( 'page_rewrite_rules', array( $this, 'rewrite_rules' ) );
+
 		\add_action( 'template_redirect', array( $this, 'text_url_response' ) );
 	}
 
@@ -164,8 +166,21 @@ class Plugin {
 	 *
 	 * Allows fetching the raw Markdown used on the page.
 	 */
-	public function text_url_permalink() {
-		\add_rewrite_rule( '.+?([^/]+?)\.text/?$', 'index.php?export=markdown&name=$matches[1]', 'top' );
+	public function rewrite_rules( $rewrite ) {
+
+		$md_rewrite = array();
+
+		foreach ( $rewrite as $pattern => $redirect ) {
+			if ( preg_match( '/name=/', $redirect ) && preg_match( '/page=/', $redirect ) ) {
+				// Repurpose the paged rewrite rules:
+				$md_pattern  = str_replace( '(/[0-9]+)?/?$', '\.text/?$', $pattern );
+				$md_redirect = preg_replace( '/([&?])page=[^\&]*/', '$1export=markdown', $redirect, 1 );
+
+				$md_rewrite[ $md_pattern ] = $md_redirect;
+			}
+		}
+
+		return array_merge( $md_rewrite, $rewrite );
 	}
 
 	/**
@@ -178,7 +193,31 @@ class Plugin {
 			&& $post->post_status === 'publish' && $post->post_password === '' ) {
 
 			header( 'Content-Type: text/markdown; charset=' . \get_bloginfo( 'charset' ) );
-			// TODO: Metadata
+			
+			$metadata = array(
+				'Title'  => \get_the_title(),
+				'Author' => \get_the_author_meta( 'display_name', $post->post_author ),
+				'Date'   => \get_the_date(),
+				'URL'    => \get_the_permalink(),
+			);
+
+			/**
+			 * Lets developers change the Markdown meta headers returned with
+			 * the document.
+			 * 
+			 * @param array    $metadata Markdown document meta headers.
+			 * @param \WP_Post $post Exported post.
+			 */
+			$metadata = \apply_filters( 'markdown_metadata', $metadata, $post );
+
+			foreach ( $metadata as $key => $value ) {
+				echo "$key: $value  \n";
+			}
+
+			if ( count( $metadata ) ) {
+				echo "\n";
+			}
+
 			echo $post->post_content;
 			exit;
 		}